Actually, time is gonna be a long

+++ Python/file/pcap/Observation.py
@@ -41,4 +41,4 @@ def __getTransmissionAddress(packet):
 
 #UNIX time in seconds
 def __getTime(packet):
-    return int(packet.time)
+    return long(packet.time)
